update UI category

+++ src/app/navigation/navbar/NavBar.tsx
@@ -15,6 +15,7 @@ import Cart from "@material-ui/icons/ShoppingCart";
 import clsx from "clsx";
 import React from "react";
 import { Link, useNavigate } from "react-router-dom";
+import { updateSwitchRole } from "../../admin/sliceSwitchRole/switchRoleSlice";
 import FooterComponent from "../../component/footer/FooterComponent";
 import { ROUTE } from "../../contant/Contant";
 import { useAppDispatch, useAppSelector } from "../../hooks";
@@ -22,24 +23,22 @@ import {
   deleteItemCart,
   updateQuantity,
 } from "../../screen/cart/slice/CartSlice";
+import { getAdmin } from "../../service/StorageService";
 import { colors } from "../../utils/color";
 import { useWindowSize } from "../../utils/helper";
-import ActiveBreadcrumbs from "../Breadcrumbs";
+import { createNotification } from "../../utils/MessageUtil";
 import MiniDrawer from "../Drawer";
 import MainApp from "../MainApp";
 import { useNavBarStyles } from "./styles";
 
 export default function NavBar() {
+  const token = localStorage.getItem("token");
   const size = useWindowSize();
   const classes = useNavBarStyles();
   const navigate = useNavigate();
   const { data } = useAppSelector((state) => state.cart);
   const dispatch = useAppDispatch();
-  const [open, setOpen] = React.useState(false);
-
-  const handleDrawerOpen = () => {
-    setOpen(true);
-  };
+  const [open, setOpen] = React.useState(true);
 
   const [anchorEl, setAnchorEl] = React.useState<null | HTMLElement>(null);
   const [mobileMoreAnchorEl, setMobileMoreAnchorEl] =
@@ -55,7 +54,15 @@ export default function NavBar() {
     setAnchorEl(event.currentTarget);
   };
   const handleCartOpen = (event: React.MouseEvent<HTMLElement>) => {
-    setCartMoreAnchorEl(event.currentTarget);
+    if (token) {
+      setCartMoreAnchorEl(event.currentTarget);
+    } else {
+      navigate(ROUTE.LOGIN);
+      createNotification({
+        type: "warning",
+        message: "Bạn cần đăng nhập để thực hiện chức năng này",
+      });
+    }
   };
 
   const handleMobileMenuClose = () => {
@@ -76,6 +83,7 @@ export default function NavBar() {
   };
 
   const menuId = "primary-search-account-menu";
+
   const renderMenu = (
     <Menu
       anchorEl={anchorEl}
@@ -86,16 +94,29 @@ export default function NavBar() {
       open={isMenuOpen}
       onClose={handleMenuClose}
     >
-      <MenuItem onClick={handleMenuClose}>Profile</MenuItem>
-      <MenuItem onClick={handleMenuClose}>My account</MenuItem>
+      {token && (
+        <>
+          <MenuItem onClick={handleMenuClose}>Profile</MenuItem>
+          <MenuItem onClick={handleMenuClose}>My account</MenuItem>
+        </>
+      )}
+
       <MenuItem
         onClick={() => {
-          localStorage.clear();
+          if (token) {
+            localStorage.clear();
+            dispatch(updateSwitchRole(false));
+
+            createNotification({
+              type: "success",
+              message: "Đăng xuất thành công",
+            });
+            handleMenuClose();
+          }
           navigate(ROUTE.LOGIN, { replace: true });
-          handleMenuClose();
         }}
       >
-        Log out
+        {token ? "Log out" : "Login app"}
       </MenuItem>
     </Menu>
   );
@@ -255,14 +276,14 @@ export default function NavBar() {
       </MenuItem>
     </Menu>
   );
-  const isAdmin = false;
+  const isAdmin = getAdmin();
   return (
     <div className={classes.grow}>
       {isAdmin ? (
         <div className={classes.containerAdmin}>
           <MiniDrawer open={open} setOpen={setOpen} />
           <div style={{ padding: 10, flex: 1 }}>
-            <MainApp isAdmin={isAdmin} />
+            <MainApp isAdmin={isAdmin ? true : false} />
           </div>
         </div>
       ) : (
@@ -278,7 +299,7 @@ export default function NavBar() {
             <AppBar
               position="fixed"
               className={clsx(classes.appBar, {
-                [classes.appBarShift]: open,
+                // [classes.appBarShift]: open,
               })}
             >
               <Toolbar>
@@ -361,7 +382,7 @@ export default function NavBar() {
                   >
                     <Cart />
                   </IconButton>
-                  {data && data?.length > 0 && (
+                  {token && data && data?.length > 0 && (
                     <div
                       style={{
                         position: "absolute",
@@ -416,7 +437,7 @@ export default function NavBar() {
               }}
             >
               {/* <ActiveBreadcrumbs /> */}
-              <MainApp isAdmin={isAdmin} />
+              <MainApp isAdmin={isAdmin ? true : false} />
             </div>
             {renderMobileMenu}
             {renderMenu}
